pyproject.toml
[project]
name = "robo-wheel-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crcmod>=1.7",
    "numpy>=2.1.1",
    "opencv-python>=4.10.0.84",
    "pyrealsense2>=2.55.1.6486",
    "pyserial>=3.5",
    "pyzmq>=26.2.0",
    "ruff>=0.6.7",
]
